Move vendor/ path setup to fjord/__init__.py.
This is in preparation for celery changes that require code to be in
this file as well and need vendor to be setup.
